Begin with Tome if Tengu has been defeated

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Tengu.java
@@ -130,7 +130,7 @@ public class Tengu extends Mob {
 	@Override
 	public void die( Object cause ) {
 		
-		if (Dungeon.hero.subClass == HeroSubClass.NONE) {
+		if (Dungeon.hero.subClass == HeroSubClass.NONE && Dungeon.hero.belongings.getItem(TomeOfMastery.class) == null) {
 			Dungeon.level.drop( new TomeOfMastery(), pos ).sprite.drop();
 		}
